requirements.txt
alabaster==0.7.12
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
auditlog3==1.0.1
autoflake==1.4
autopep8==1.6.0
Babel==2.9.1
bandit==1.7.4
black==22.1.0
captcha==0.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
codecov==2.1.12
colorama==0.4.4
coverage==6.3.2
cryptography==36.0.1
Deprecated==1.2.13
Django==3.2.12
django-appconf==1.0.5
django-classy-tags==3.0.1
django-cms==3.9.0
django-email-extras==0.3.4
django-form-builder==0.16.7
django-forms-builder==0.14.0
django-formtools==2.3
django-ipware==4.0.2
django-sekizai==3.0.1
django-treebeard==4.5.1
djangocms-admin-style==3.0.0
djangocms-forms==0.2.5
djangorestframework==3.13.1
docutils==0.17.1
filesig==0.3.0.post1
flake8==2.3.0
future==0.15.0
gcloud==0.18.3
gitdb==4.0.9
GitPython==3.1.27
googleapis-common-protos==1.55.0
gunicorn==20.1.0
hashids==1.3.1
httplib2==0.20.4
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.2
isort==5.10.1
Jinja2==3.0.3
jsonfield==3.1.0
jwcrypto==1.0
MarkupSafe==2.1.0
mccabe==0.7.0
more-itertools==8.12.0
mypy-extensions==0.4.3
oauth2client==4.1.3
packaging==21.3
pathspec==0.9.0
pbr==5.8.1
pep8==1.7.1
Pillow==9.0.1
platformdirs==2.5.1
pluggy==0.13.1
protobuf==3.20.0rc1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.14.1
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.11.2
pylama==8.3.8
pyparsing==3.0.7
Pyrebase4==4.5.0
pytest==5.4.1
pytest-cov==3.0.0
pytest-django==3.8.0
pytest-pythonpath==0.7.3
python-dateutil==2.7.5
python-gnupg==0.4.8
python-jwt==3.3.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==4.4.0
sphinx-me==0.3
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-images==0.9.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
stevedore==3.5.0
tablib==3.2.0
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.2
typing_extensions==4.1.1
Unidecode==1.3.4
urllib3==1.26.8
wcwidth==0.2.5
wrapt==1.14.0
zipp==3.7.0
